10_lesson/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
allure-pytest==2.13.2
pytest==7.4.3
flake8==6.1.0